Clean up testOnly issue (?) and fix up some state tracking issues

+++ app/src/main/java/com/androidexperiments/lipflip/utils/FileUtils.java
@@ -511,10 +511,11 @@ public class FileUtils {
                 + File.separator;
         File directory = new File(path);
         File[] files = directory.listFiles();
-
-        for (int i = 0; i < files.length; i++) {
-            if (files[i].length() == 0) {
-                files[i].delete();
+        if (files != null) {
+            for (int i = 0; i < files.length; i++) {
+                if (files[i].length() == 0) {
+                    files[i].delete();
+                }
             }
         }
